Project 474/Streaming-Data-Pipeline/producer/requirements.txt
# === Kafka Client ===
kafka-python==2.0.2

# === Networking / Requests ===
requests==2.31.0